ASPSH-1018: Setup doorkeeper and devise for oauth
Initial POC of how oauth will look for clients that want access to a user's
data.

Devise will be used for authentication and doorkeeper for authorization.
Authorization code grant flow will the primary way a client can get an API key
to access a resource owner's data.

To get access to the messaging queue "paperboy" requires a different
authentication method. This will be done in another ticket.